requirements.txt
torch
diffusers
transformers
Pillow
pytest